+++ source/core/gameinput.h
@@ -32,7 +32,7 @@ class GameInput
 	ESyncBits ActionsToSend;
 
 	// Turn speed doubling after x amount of tics.
-	void updateTurnHeldAmt(const double scaleAdjust)
+	void updateTurnHeldAmt(const float scaleAdjust)
 	{
 		turnheldtime += getTicrateScale(BUILDTICRATE) * scaleAdjust;
 	}
@@ -75,8 +75,8 @@ public:
 	}
 
 	// Prototypes for large member functions.
-	void processMovement(const double scaleAdjust, const int drink_amt = 0, const bool allowstrafe = true, const double turnscale = 1.);
-	void processVehicle(const double scaleAdjust, const float baseVel, const float velScale, const bool canMove, const bool canTurn, const bool attenuate);
+	void processMovement(const float scaleAdjust, const int drink_amt = 0, const bool allowstrafe = true, const double turnscale = 1.);
+	void processVehicle(const float scaleAdjust, const float baseVel, const float velScale, const bool canMove, const bool canTurn, const bool attenuate);
 	void getInput(const double scaleAdjust, InputPacket* packet = nullptr);
 };
 
@@ -89,8 +89,8 @@ struct PlayerAngles
 	DAngle YawSpin;
 
 	friend FSerializer& Serialize(FSerializer& arc, const char* keyname, PlayerAngles& w, PlayerAngles* def);
-	friend void GameInput::processMovement(const double scaleAdjust, const int drink_amt, const bool allowstrafe, const double turnscale);
-	friend void GameInput::processVehicle(const double scaleAdjust, const float baseVel, const float velScale, const bool canMove, const bool canTurn, const bool attenuate);
+	friend void GameInput::processMovement(const float scaleAdjust, const int drink_amt, const bool allowstrafe, const double turnscale);
+	friend void GameInput::processVehicle(const float scaleAdjust, const float baseVel, const float velScale, const bool canMove, const bool canTurn, const bool attenuate);
 
 	// Prototypes.
 	void doPitchKeys(InputPacket* const input);
